requirements.txt
celery==5.0.5
Django==3.2.4
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-guardian==2.4.0
djangorestframework==3.12.4
#flower==0.9.7 # show celery info
Markdown==3.3.4 # browsable API
Pillow==8.2.0 # django ImageField
redis==3.5.3
drf-yasg==1.20.0 #api documentation
coverage==5.5 #API test coverage
gunicorn==20.1.0